Histogram: Correlation, Chi-Squared, Intersection, Hellinger
This, in turn, was the best scoring technique used and i think it would be viable for use in a program. I found through much trial and error that the "Correlation" method gave the best results. You can see from my previous printouts that this method gave me the best scores.