Block Search Algorithms for Tracking
Searches include: 

HexagonBM- Use the center of in each round to find the best value.
Diamond- Use the center of in each round to find the best value.
4SS (4 Stage Search) - Use the center of in each round to find the best value.
3SS (3 Stage Search) - Use the center of in each round to find the best value.
Full Search - Use the center of in each round to find the best value in a 7x7 area surrounding query point. Code Full Search as a BxB location (assume it can be centered by B being odd valued).